Role based Authorization added

+++ src/main/java/com/example/journalApp/service/JournalEntryService.java
@@ -27,7 +27,7 @@ public class JournalEntryService {
             entry.setDate(LocalDateTime.now());
             JournalEntry saved = journalRepo.save(entry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveNewUser(user);
         }
         catch(Exception e){
             System.out.println(e);
@@ -50,10 +50,20 @@ public class JournalEntryService {
         return journalRepo.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalRepo.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed) {
+                userService.saveUser(user);
+                journalRepo.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry", e);
+        }
+        return removed;
     }
 }
